// src/oem/main/oem-notifications/oem-notification.enums/oem-notification.notification-type.enum.ts
export enum OemNotificationTypeEnum {
  VENDO_APPROVED = 'Vendo Approved',
  VENDO_REJECTED = 'Vendo Rejected',
  VENDO_SUBMITTED = 'Vendo Submitted',
  VENDO_CHANGED = 'Vendo Changed',
  VENDO_EXPIRED = 'Vendo Expired',
  VENDO_TRANSACTED = 'Vendo Transacted',

  QUOTE_APPROVED = 'Quote Approved',
  QUOTE_REJECTED = 'Quote Rejected',
  QUOTE_SUBMITTED = 'Quote Submitted',
  QUOTE_CHANGED = 'Quote Changed',
  QUOTE_EXPIRED = 'Quote Expired',
  QUOTE_TRANSACTED = 'Quote Transacted',

  CUSTOM_ALERT = 'Custom Alert',
  BATCHED_USER_EMAIL = 'Batched User Email',
  USER_VERIFY = 'User Verify',
  USER_INVITE = 'User Invite',
  FORGOT_PASSWORD = 'Forgot Password',
}